workflow: remove public schedule and trigger CRUD
Schedules and triggers are now YAML-only:

    workflows:
      schedules:
        nightly_sync:
          provider: local
          cron: "0 3 * * *"
          target:
            plugin:
              name: roadmap
              operation: sync

      eventTriggers:
        item_updated:
          provider: local
          match:
            type: roadmap.item.updated
          target:
            plugin:
              name: roadmap
              operation: sync

The public workflow API/CLI now keeps event publishing and run inspection/control:

    POST /api/v1/workflow/events
    GET /api/v1/workflow/runs
    GET /api/v1/workflow/runs/{runID}
    POST /api/v1/workflow/runs/{runID}/cancel

    gestalt workflow events publish --type roadmap.item.updated
    gestalt workflow runs list
    gestalt workflow runs get <run-id>
    gestalt workflow runs cancel <run-id> --reason "operator requested"

This removes the public workflow schedule/event-trigger HTTP handlers, CLI subcommands, and agent workflow schedule system tools while preserving config-managed cron schedules and event triggers.
