+++ docs/content/client/web.mdx
@@ -18,9 +18,8 @@ and workflow counts.
 When the default web bundle is mounted, the workflow UI is available at
 `/workflows`.
 
-It shows workflow schedules, event triggers, and recent run history in one
-place. You can inspect individual runs, filter by status, search recent
-activity, and cancel pending runs.
+It shows recent workflow run history. You can inspect individual runs, filter
+by status, search recent activity, and cancel pending runs.
 
 ## Plugins
