+++ docs/content/configuration.mdx
@@ -116,7 +116,7 @@ Gestalt has ten core concepts you configure in YAML:
 - **[S3.](/providers/s3)** Plugin-bound object storage. The provider config chooses the backend, credentials, and endpoint; plugin code chooses buckets, keys, and versions at runtime.
 - **[Authentication.](/providers/authentication)** Platform login for the Gestalt server. Separate from connection authentication, which is per-plugin and per-upstream.
 - **[Secrets.](/providers/secrets)** Resolves structured secret refs in the config at startup. Backed by environment variables, files, or cloud secret managers.
-- **[Workflow.](/providers/workflow)** Global workflow runs, schedules, and triggers backed by a workflow provider and optional host IndexedDB storage.
+- **[Workflow.](/providers/workflow)** Global workflow runs plus config-managed schedules and triggers backed by a workflow provider and optional host IndexedDB storage.
 - **[UI.](/providers/ui)** Public static UI bundles served under configured path prefixes, either directly from `providers.ui` or through `plugins.<name>.ui.path`. Omit `providers.ui` to run headless.
 - **Runtime.** Optional execution backends for executable plugins. Plugins stay host-local by default; `plugins.<name>.execution.mode: hosted` opts a plugin into hosted execution and `runtime.providers` names the available backends.
 
@@ -510,9 +510,7 @@ and the built-in admin authorization APIs.
 The `providers.workflow` block declares workflow backends. Config-managed
 entries under top-level `workflows.schedules` and `workflows.eventTriggers`
 select a provider with `provider`, or inherit the sole/default
-`providers.workflow` entry when that field is omitted. User-owned schedules use
-the same provider pool through the global workflow API and CLI surface:
-`/api/v1/workflow/schedules` and `gestalt workflow ...`.
+`providers.workflow` entry when that field is omitted.
 
 ```yaml
 providers:
