+++ docs/content/reference/config-file.mdx
@@ -421,9 +421,7 @@ top-level `workflows.schedules.*.provider` or
 `workflows.eventTriggers.*.provider`. There is no
 `server.providers.workflow`: config-managed workflow objects choose a provider
 explicitly, or inherit the sole/default workflow provider when `provider` is
-omitted. User-owned schedules and triggers use the same provider pool
-through `/api/v1/workflow/schedules`,
-`/api/v1/workflow/event-triggers`, and `gestalt workflow ...`.
+omitted.
 
 ```yaml
 providers:
@@ -474,7 +472,7 @@ allowlist logical stores exposed through that binding.
 
 | Field | Description |
 | --- | --- |
-| `default` | Marks this named workflow entry as the default selection when a config-managed workflow object or API request omits `provider`. |
+| `default` | Marks this named workflow entry as the default selection when a config-managed workflow object omits `provider`. |
 | `source` | Either a local provider manifest path, a published `provider-release.yaml` metadata URL, or `source.githubRelease` for a GitHub-hosted release asset. |
 | `auth` | Optional inline source auth for remote release sources, including direct metadata URLs and `source.githubRelease`. |
 | `config` | Provider-specific config passed into the workflow provider. |
@@ -556,9 +554,7 @@ invoke an explicit `target`. A target is exactly one of `target.plugin` or
 
 Config-managed workflows are deployment-owned. They are created from YAML at
 startup, updated when config changes, and removed when deleted from config.
-They are distinct from user-owned schedules created through the global workflow
-HTTP API and CLI. For an end-to-end usage guide, see
-[Workflow](/providers/workflow).
+For an end-to-end usage guide, see [Workflow](/providers/workflow).
 
 ```yaml
 workflows:
@@ -624,9 +620,7 @@ workflows:
 | `paused` | When `true`, Gestalt reconciles the trigger in a paused state. |
 
 Event trigger matching is exact string matching. `match.type` is required;
-`match.source` and `match.subject` are optional exact-match filters. Event
-triggers can be config-managed through YAML or user-owned through the global
-workflow HTTP API and CLI.
+`match.source` and `match.subject` are optional exact-match filters.
 
 ## `providers.cache`
