+++ gestalt/cli/tests/workflows_cli.rs
@@ -2,40 +2,6 @@ mod support;
 
 use support::*;
 
-const SCHEDULE_JSON: &str = r#"{
-    "id":"sched-1",
-    "provider":"test-provider",
-    "cron":"0 0 * * *",
-    "timezone":"UTC",
-    "target":{
-        "plugin":{
-            "name":"dummy",
-            "operation":"doit",
-            "input":{"k":"v"}
-        }
-    },
-    "paused":false,
-    "createdAt":"2026-04-20T00:00:00Z",
-    "updatedAt":"2026-04-20T00:00:00Z",
-    "nextRunAt":"2026-04-21T00:00:00Z"
-}"#;
-
-const TRIGGER_JSON: &str = r#"{
-    "id":"trg-1",
-    "provider":"test-provider",
-    "match":{"type":"dummy.event","source":"dummy","subject":"item"},
-    "target":{
-        "plugin":{
-            "name":"dummy",
-            "operation":"doit",
-            "input":{"k":"v"}
-        }
-    },
-    "paused":false,
-    "createdAt":"2026-04-20T00:00:00Z",
-    "updatedAt":"2026-04-20T00:00:00Z"
-}"#;
-
 const RUN_JSON: &str = r#"{
     "id":"run-1",
     "provider":"test-provider",
@@ -68,536 +34,6 @@ const PUBLISHED_EVENT_JSON: &str = r#"{
         "extensions":{"traceId":"trace-1"}
     }
 }"#;
-
-#[test]
-fn test_cli_lists_schedules() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules",
-        StatusCode::OK
-    )
-    .with_body(format!("[{SCHEDULE_JSON}]"))
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "schedules", "list"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-1"))
-        .stdout(predicate::str::contains("dummy"))
-        .stdout(predicate::str::contains("doit"));
-
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules",
-        StatusCode::OK
-    )
-    .with_body(format!("[{SCHEDULE_JSON}]"))
-    .create();
-
-    cli_command_for_server(home.path(), &server)
-        .args(["workflows", "schedules", "list"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-1"));
-}
-
-#[test]
-fn test_cli_list_schedules_filters_by_plugin() {
-    let body = r#"[
-        {"id":"sched-a","provider":"p","cron":"* * * * *","target":{"plugin":{"name":"alpha","operation":"x"}},"paused":false},
-        {"id":"sched-b","provider":"p","cron":"* * * * *","target":{"plugin":{"name":"beta","operation":"y"}},"paused":false}
-    ]"#;
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules",
-        StatusCode::OK
-    )
-    .with_body(body)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "schedules", "list", "--plugin", "beta"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-b"))
-        .stdout(predicate::str::contains("sched-a").not());
-}
-
-#[test]
-fn test_cli_gets_schedule() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules/sched-1",
-        StatusCode::OK
-    )
-    .with_body(SCHEDULE_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "schedules", "get", "sched-1"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-1"))
-        .stdout(predicate::str::contains("dummy"));
-}
-
-#[test]
-fn test_cli_creates_schedule() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::POST,
-        "/api/v1/workflow/schedules",
-        StatusCode::CREATED
-    )
-    .match_header(header::CONTENT_TYPE.as_str(), http::APPLICATION_JSON)
-    .match_body(Matcher::JsonString(
-        r#"{
-            "cron":"0 */5 * * *",
-            "timezone":"UTC",
-            "target":{"plugin":{"name":"dummy","operation":"doit","input":{"channel":"C1","text":"hi"}}},
-            "paused":false
-        }"#
-        .to_string(),
-    ))
-    .with_body(SCHEDULE_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args([
-            "workflow",
-            "schedules",
-            "create",
-            "--cron",
-            "0 */5 * * *",
-            "--timezone",
-            "UTC",
-            "--plugin",
-            "dummy",
-            "--operation",
-            "doit",
-            "-p",
-            "channel=C1",
-            "-p",
-            "text=hi",
-        ])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-1"));
-}
-
-#[test]
-fn test_cli_updates_schedule_merges_existing_fields() {
-    let mut server = Server::new();
-    let _get = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules/sched-1",
-        StatusCode::OK
-    )
-    .with_body(SCHEDULE_JSON)
-    .create();
-
-    let _put = authed_json_mock!(
-        server,
-        Method::PUT,
-        "/api/v1/workflow/schedules/sched-1",
-        StatusCode::OK
-    )
-    .match_header(header::CONTENT_TYPE.as_str(), http::APPLICATION_JSON)
-    .match_body(Matcher::JsonString(
-        r#"{
-            "cron":"15 * * * *",
-            "timezone":"UTC",
-            "target":{"plugin":{"name":"dummy","operation":"doit","input":{"k":"v"}}},
-            "paused":true
-        }"#
-        .to_string(),
-    ))
-    .with_body(SCHEDULE_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args([
-            "workflow",
-            "schedules",
-            "update",
-            "sched-1",
-            "--cron",
-            "15 * * * *",
-            "--paused",
-        ])
-        .assert()
-        .success();
-}
-
-#[test]
-fn test_cli_updates_schedule_requires_canonical_existing_target() {
-    let legacy_schedule = r#"{
-        "id":"sched-legacy",
-        "provider":"test-provider",
-        "cron":"0 0 * * *",
-        "timezone":"UTC",
-        "target":{
-            "plugin":"legacy",
-            "operation":"sync",
-            "connection":"analytics",
-            "instance":"tenant-a",
-            "input":{"k":"v"}
-        },
-        "paused":false
-    }"#;
-    let mut server = Server::new();
-    let _get = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules/sched-legacy",
-        StatusCode::OK
-    )
-    .with_body(legacy_schedule)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args([
-            "workflow",
-            "schedules",
-            "update",
-            "sched-legacy",
-            "--cron",
-            "15 * * * *",
-        ])
-        .assert()
-        .failure()
-        .stderr(predicate::str::contains(
-            "existing schedule is missing target.plugin.name; pass --plugin",
-        ));
-}
-
-#[test]
-fn test_cli_deletes_schedule() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::DELETE,
-        "/api/v1/workflow/schedules/sched-1",
-        StatusCode::OK
-    )
-    .with_body(r#"{"status":"deleted"}"#)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "schedules", "delete", "sched-1"])
-        .assert()
-        .success()
-        .stderr(predicate::str::contains(
-            "Workflow schedule sched-1 deleted.",
-        ));
-}
-
-#[test]
-fn test_cli_pauses_schedule() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::POST,
-        "/api/v1/workflow/schedules/sched-1/pause",
-        StatusCode::OK
-    )
-    .with_body(SCHEDULE_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "schedules", "pause", "sched-1"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-1"));
-}
-
-#[test]
-fn test_cli_resumes_schedule() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::POST,
-        "/api/v1/workflow/schedules/sched-1/resume",
-        StatusCode::OK
-    )
-    .with_body(SCHEDULE_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "schedules", "resume", "sched-1"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("sched-1"));
-}
-
-#[test]
-fn test_cli_list_schedules_json_format() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/schedules",
-        StatusCode::OK
-    )
-    .with_body(format!("[{SCHEDULE_JSON}]"))
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["--format", "json", "workflow", "schedules", "list"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains(r#""id": "sched-1""#))
-        .stdout(predicate::str::contains(r#""name": "dummy""#));
-}
-
-#[test]
-fn test_cli_lists_event_triggers() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/event-triggers",
-        StatusCode::OK
-    )
-    .with_body(format!("[{TRIGGER_JSON}]"))
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "triggers", "list"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("trg-1"))
-        .stdout(predicate::str::contains("dummy"))
-        .stdout(predicate::str::contains("doit"));
-}
-
-#[test]
-fn test_cli_list_event_triggers_filters() {
-    let body = r#"[
-        {"id":"trg-a","match":{"type":"alpha.created"},"target":{"plugin":{"name":"alpha","operation":"x"}},"paused":false},
-        {"id":"trg-b","match":{"type":"beta.failed"},"target":{"plugin":{"name":"beta","operation":"y"}},"paused":false}
-    ]"#;
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/event-triggers",
-        StatusCode::OK
-    )
-    .with_body(body)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args([
-            "workflow",
-            "triggers",
-            "list",
-            "--plugin",
-            "beta",
-            "--type",
-            "beta.failed",
-        ])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("trg-b"))
-        .stdout(predicate::str::contains("trg-a").not());
-}
-
-#[test]
-fn test_cli_gets_event_trigger() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/event-triggers/trg-1",
-        StatusCode::OK
-    )
-    .with_body(TRIGGER_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "triggers", "get", "trg-1"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("trg-1"))
-        .stdout(predicate::str::contains("dummy"));
-}
-
-#[test]
-fn test_cli_creates_event_trigger() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::POST,
-        "/api/v1/workflow/event-triggers",
-        StatusCode::CREATED
-    )
-    .match_header(header::CONTENT_TYPE.as_str(), http::APPLICATION_JSON)
-    .match_body(Matcher::JsonString(
-        r#"{
-            "match":{"type":"dummy.event","source":"dummy","subject":"item"},
-            "target":{"plugin":{"name":"dummy","operation":"doit","input":{"channel":"C1","text":"hi"}}},
-            "paused":false
-        }"#
-        .to_string(),
-    ))
-    .with_body(TRIGGER_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args([
-            "workflow",
-            "triggers",
-            "create",
-            "--type",
-            "dummy.event",
-            "--source",
-            "dummy",
-            "--subject",
-            "item",
-            "--plugin",
-            "dummy",
-            "--operation",
-            "doit",
-            "-p",
-            "channel=C1",
-            "-p",
-            "text=hi",
-        ])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("trg-1"));
-}
-
-#[test]
-fn test_cli_updates_event_trigger_merges_existing_fields() {
-    let mut server = Server::new();
-    let _get = authed_json_mock!(
-        server,
-        Method::GET,
-        "/api/v1/workflow/event-triggers/trg-1",
-        StatusCode::OK
-    )
-    .with_body(TRIGGER_JSON)
-    .create();
-
-    let _put = authed_json_mock!(
-        server,
-        Method::PUT,
-        "/api/v1/workflow/event-triggers/trg-1",
-        StatusCode::OK
-    )
-    .match_header(header::CONTENT_TYPE.as_str(), http::APPLICATION_JSON)
-    .match_body(Matcher::JsonString(
-        r#"{
-            "provider":"test-provider",
-            "match":{"type":"dummy.event.updated","source":"dummy","subject":"item"},
-            "target":{"plugin":{"name":"dummy","operation":"doit","input":{"k":"v"}}},
-            "paused":true
-        }"#
-        .to_string(),
-    ))
-    .with_body(TRIGGER_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args([
-            "workflow",
-            "triggers",
-            "update",
-            "trg-1",
-            "--type",
-            "dummy.event.updated",
-            "--paused",
-        ])
-        .assert()
-        .success();
-}
-
-#[test]
-fn test_cli_deletes_event_trigger() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::DELETE,
-        "/api/v1/workflow/event-triggers/trg-1",
-        StatusCode::OK
-    )
-    .with_body(r#"{"status":"deleted"}"#)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "triggers", "delete", "trg-1"])
-        .assert()
-        .success()
-        .stderr(predicate::str::contains("Workflow trigger trg-1 deleted."));
-}
-
-#[test]
-fn test_cli_pauses_event_trigger() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::POST,
-        "/api/v1/workflow/event-triggers/trg-1/pause",
-        StatusCode::OK
-    )
-    .with_body(TRIGGER_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "triggers", "pause", "trg-1"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("trg-1"));
-}
-
-#[test]
-fn test_cli_resumes_event_trigger() {
-    let mut server = Server::new();
-    let _mock = authed_json_mock!(
-        server,
-        Method::POST,
-        "/api/v1/workflow/event-triggers/trg-1/resume",
-        StatusCode::OK
-    )
-    .with_body(TRIGGER_JSON)
-    .create();
-
-    let home = tempfile::tempdir().unwrap();
-    cli_command_for_server(home.path(), &server)
-        .args(["workflow", "triggers", "resume", "trg-1"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("trg-1"));
-}
 
 #[test]
 fn test_cli_lists_runs() {
