+++ gestaltd/internal/bootstrap/agent_workflow_tools.go
@@ -10,24 +10,17 @@ import (
 	"strings"
 	"time"
 
-	"github.com/valon-technologies/gestalt/server/core"
 	coreagent "github.com/valon-technologies/gestalt/server/core/agent"
 	coreworkflow "github.com/valon-technologies/gestalt/server/core/workflow"
 	"github.com/valon-technologies/gestalt/server/internal/agentmanager"
-	"github.com/valon-technologies/gestalt/server/internal/config"
 	"github.com/valon-technologies/gestalt/server/internal/invocation"
 	"github.com/valon-technologies/gestalt/server/internal/principal"
 	"github.com/valon-technologies/gestalt/server/internal/workflowmanager"
 )
 
 const (
-	workflowSystemToolSchedulesCreate = "schedules.create"
-	workflowSystemToolSchedulesList   = "schedules.list"
-	workflowSystemToolSchedulesGet    = "schedules.get"
-	workflowSystemToolSchedulesPause  = "schedules.pause"
-	workflowSystemToolSchedulesResume = "schedules.resume"
-	workflowSystemToolRunsList        = "runs.list"
-	workflowSystemToolRunsGet         = "runs.get"
+	workflowSystemToolRunsList = "runs.list"
+	workflowSystemToolRunsGet  = "runs.get"
 )
 
 type workflowSystemToolAvailability interface {
@@ -51,36 +44,6 @@ func newWorkflowSystemTools(manager workflowmanager.Service, availability workfl
 }
 
 var workflowSystemToolDescriptors = map[string]workflowSystemToolDescriptor{
-	workflowSystemToolSchedulesCreate: {
-		Operation:        workflowSystemToolSchedulesCreate,
-		Name:             "workflow_schedules_create",
-		Description:      "Create a recurring workflow schedule for a delegated target.",
-		ParametersSchema: workflowSystemToolCreateScheduleSchema(),
-	},
-	workflowSystemToolSchedulesList: {
-		Operation:        workflowSystemToolSchedulesList,
-		Name:             "workflow_schedules_list",
-		Description:      "List workflow schedules owned by the current caller.",
-		ParametersSchema: workflowSystemToolObjectSchema(nil, nil),
-	},
-	workflowSystemToolSchedulesGet: {
-		Operation:        workflowSystemToolSchedulesGet,
-		Name:             "workflow_schedules_get",
-		Description:      "Get a workflow schedule owned by the current caller.",
-		ParametersSchema: workflowSystemToolObjectSchema([]string{"scheduleId"}, map[string]any{"scheduleId": workflowSystemToolStringSchema("Schedule ID.")}),
-	},
-	workflowSystemToolSchedulesPause: {
-		Operation:        workflowSystemToolSchedulesPause,
-		Name:             "workflow_schedules_pause",
-		Description:      "Pause a workflow schedule owned by the current caller.",
-		ParametersSchema: workflowSystemToolObjectSchema([]string{"scheduleId"}, map[string]any{"scheduleId": workflowSystemToolStringSchema("Schedule ID.")}),
-	},
-	workflowSystemToolSchedulesResume: {
-		Operation:        workflowSystemToolSchedulesResume,
-		Name:             "workflow_schedules_resume",
-		Description:      "Resume a workflow schedule owned by the current caller.",
-		ParametersSchema: workflowSystemToolObjectSchema([]string{"scheduleId"}, map[string]any{"scheduleId": workflowSystemToolStringSchema("Schedule ID.")}),
-	},
 	workflowSystemToolRunsList: {
 		Operation:        workflowSystemToolRunsList,
 		Name:             "workflow_runs_list",
@@ -148,60 +111,6 @@ func (t *workflowSystemTools) ExecuteSystemTool(ctx context.Context, req agentSy
 		return nil, fmt.Errorf("%w: unsupported agent system tool %q", invocation.ErrInvalidInvocation, req.Tool.Target.System)
 	}
 	switch strings.TrimSpace(req.Tool.Target.Operation) {
-	case workflowSystemToolSchedulesCreate:
-		return t.executeCreateSchedule(ctx, req)
-	case workflowSystemToolSchedulesList:
-		if err := workflowSystemToolRejectUnknownKeys(req.Arguments, "workflow.schedules.list"); err != nil {
-			return nil, err
-		}
-		schedules, err := t.manager.ListSchedules(ctx, req.Principal)
-		if err != nil {
-			return nil, err
-		}
-		items := make([]map[string]any, 0, len(schedules))
-		for _, schedule := range schedules {
-			items = append(items, workflowSystemToolScheduleInfo(schedule))
-		}
-		return workflowSystemToolJSONResponse(http.StatusOK, map[string]any{"schedules": items})
-	case workflowSystemToolSchedulesGet:
-		if err := workflowSystemToolRejectUnknownKeys(req.Arguments, "workflow.schedules.get", "scheduleId", "id"); err != nil {
-			return nil, err
-		}
-		scheduleID := workflowSystemToolStringArg(req.Arguments, "scheduleId", "id")
-		if scheduleID == "" {
-			return nil, fmt.Errorf("%w: scheduleId is required", invocation.ErrInvalidInvocation)
-		}
-		schedule, err := t.manager.GetSchedule(ctx, req.Principal, scheduleID)
-		if err != nil {
-			return nil, err
-		}
-		return workflowSystemToolJSONResponse(http.StatusOK, map[string]any{"schedule": workflowSystemToolScheduleInfo(schedule)})
-	case workflowSystemToolSchedulesPause:
-		if err := workflowSystemToolRejectUnknownKeys(req.Arguments, "workflow.schedules.pause", "scheduleId", "id"); err != nil {
-			return nil, err
-		}
-		scheduleID := workflowSystemToolStringArg(req.Arguments, "scheduleId", "id")
-		if scheduleID == "" {
-			return nil, fmt.Errorf("%w: scheduleId is required", invocation.ErrInvalidInvocation)
-		}
-		schedule, err := t.manager.PauseSchedule(ctx, req.Principal, scheduleID)
-		if err != nil {
-			return nil, err
-		}
-		return workflowSystemToolJSONResponse(http.StatusOK, map[string]any{"schedule": workflowSystemToolScheduleInfo(schedule)})
-	case workflowSystemToolSchedulesResume:
-		if err := workflowSystemToolRejectUnknownKeys(req.Arguments, "workflow.schedules.resume", "scheduleId", "id"); err != nil {
-			return nil, err
-		}
-		scheduleID := workflowSystemToolStringArg(req.Arguments, "scheduleId", "id")
-		if scheduleID == "" {
-			return nil, fmt.Errorf("%w: scheduleId is required", invocation.ErrInvalidInvocation)
-		}
-		schedule, err := t.manager.ResumeSchedule(ctx, req.Principal, scheduleID)
-		if err != nil {
-			return nil, err
-		}
-		return workflowSystemToolJSONResponse(http.StatusOK, map[string]any{"schedule": workflowSystemToolScheduleInfo(schedule)})
 	case workflowSystemToolRunsList:
 		if err := workflowSystemToolRejectUnknownKeys(req.Arguments, "workflow.runs.list"); err != nil {
 			return nil, err
@@ -231,44 +140,6 @@ func (t *workflowSystemTools) ExecuteSystemTool(ctx context.Context, req agentSy
 	default:
 		return nil, fmt.Errorf("%w: workflow system operation %q is not supported", invocation.ErrOperationNotFound, req.Tool.Target.Operation)
 	}
-}
-
-func (t *workflowSystemTools) executeCreateSchedule(ctx context.Context, req agentSystemToolExecutionRequest) (*coreagent.ExecuteToolResponse, error) {
-	args := req.Arguments
-	if err := workflowSystemToolRejectUnknownKeys(args, "workflow.schedules.create", "provider", "providerName", "cron", "timezone", "timeZone", "paused", "target"); err != nil {
-		return nil, err
-	}
-	cron := workflowSystemToolStringArg(args, "cron")
-	if cron == "" {
-		return nil, fmt.Errorf("%w: cron is required", invocation.ErrInvalidInvocation)
-	}
-	targetValue, ok := args["target"]
-	if !ok {
-		return nil, fmt.Errorf("%w: target is required", invocation.ErrInvalidInvocation)
-	}
-	target, err := workflowSystemToolTargetFromValue(targetValue)
-	if err != nil {
-		return nil, err
-	}
-	if err := workflowSystemToolValidateCreateScope(req, target); err != nil {
-		return nil, err
-	}
-	permissions := workflowSystemToolPermissionsForTarget(target)
-	scopedPrincipal, err := workflowSystemToolScopedPrincipal(req.Principal, permissions)
-	if err != nil {
-		return nil, err
-	}
-	schedule, err := t.manager.CreateSchedule(ctx, scopedPrincipal, workflowmanager.ScheduleUpsert{
-		ProviderName: workflowSystemToolStringArg(args, "provider", "providerName"),
-		Cron:         cron,
-		Timezone:     workflowSystemToolStringArg(args, "timezone", "timeZone"),
-		Target:       target,
-		Paused:       workflowSystemToolBoolArg(args, "paused"),
-	})
-	if err != nil {
-		return nil, err
-	}
-	return workflowSystemToolJSONResponse(http.StatusCreated, map[string]any{"schedule": workflowSystemToolScheduleInfo(schedule)})
 }
 
 func workflowSystemToolFromRef(ref coreagent.ToolRef) (coreagent.Tool, error) {
@@ -301,37 +172,6 @@ func workflowSystemToolFromRef(ref coreagent.ToolRef) (coreagent.Tool, error) {
 	}, nil
 }
 
-func workflowSystemToolCreateScheduleSchema() map[string]any {
-	targetSchema := workflowSystemToolObjectSchema([]string{}, map[string]any{
-		"plugin": workflowSystemToolObjectSchema([]string{"name", "operation"}, map[string]any{
-			"name":       workflowSystemToolStringSchema("Plugin name."),
-			"operation":  workflowSystemToolStringSchema("Plugin operation."),
-			"connection": workflowSystemToolStringSchema("Connection name."),
-			"instance":   workflowSystemToolStringSchema("Instance name."),
-			"input":      map[string]any{"type": "object"},
-		}),
-		"agent": workflowSystemToolObjectSchema([]string{}, map[string]any{
-			"provider":        workflowSystemToolStringSchema("Agent provider name."),
-			"model":           workflowSystemToolStringSchema("Agent model."),
-			"prompt":          workflowSystemToolStringSchema("Agent prompt."),
-			"messages":        map[string]any{"type": "array", "items": map[string]any{"type": "object"}},
-			"toolRefs":        map[string]any{"type": "array", "items": map[string]any{"type": "object"}},
-			"tools":           map[string]any{"type": "array", "items": map[string]any{"type": "object"}},
-			"responseSchema":  map[string]any{"type": "object"},
-			"metadata":        map[string]any{"type": "object"},
-			"providerOptions": map[string]any{"type": "object"},
-			"timeoutSeconds":  map[string]any{"type": "integer", "minimum": 0},
-		}),
-	})
-	return workflowSystemToolObjectSchema([]string{"cron", "target"}, map[string]any{
-		"provider": workflowSystemToolStringSchema("Workflow provider name."),
-		"cron":     workflowSystemToolStringSchema("Cron expression."),
-		"timezone": workflowSystemToolStringSchema("IANA timezone."),
-		"paused":   map[string]any{"type": "boolean"},
-		"target":   targetSchema,
-	})
-}
-
 func workflowSystemToolObjectSchema(required []string, properties map[string]any) map[string]any {
 	if properties == nil {
 		properties = map[string]any{}
@@ -361,352 +201,6 @@ func workflowSystemToolJSONResponse(status int, value any) (*coreagent.ExecuteTo
 		return nil, fmt.Errorf("%w: marshal workflow tool response: %v", invocation.ErrInternal, err)
 	}
 	return &coreagent.ExecuteToolResponse{Status: status, Body: string(body)}, nil
-}
-
-func workflowSystemToolTargetFromValue(value any) (coreworkflow.Target, error) {
-	target, ok := workflowSystemToolMap(value)
-	if !ok {
-		return coreworkflow.Target{}, fmt.Errorf("%w: target must be an object", invocation.ErrInvalidInvocation)
-	}
-	if err := workflowSystemToolRejectUnknownKeys(target, "target", "plugin", "agent"); err != nil {
-		return coreworkflow.Target{}, err
-	}
-	pluginValue, hasPlugin := target["plugin"]
-	agentValue, hasAgent := target["agent"]
-	if hasPlugin == hasAgent {
-		return coreworkflow.Target{}, fmt.Errorf("%w: target must set exactly one of plugin or agent", invocation.ErrInvalidInvocation)
-	}
-	if hasPlugin {
-		pluginMap, ok := workflowSystemToolMap(pluginValue)
-		if !ok {
-			return coreworkflow.Target{}, fmt.Errorf("%w: target.plugin must be an object", invocation.ErrInvalidInvocation)
-		}
-		if err := workflowSystemToolRejectUnknownKeys(pluginMap, "target.plugin", "name", "plugin", "operation", "connection", "instance", "input"); err != nil {
-			return coreworkflow.Target{}, err
-		}
-		pluginName := workflowSystemToolStringArg(pluginMap, "name", "plugin")
-		operation := workflowSystemToolStringArg(pluginMap, "operation")
-		if pluginName == "" || operation == "" {
-			return coreworkflow.Target{}, fmt.Errorf("%w: target.plugin.name and target.plugin.operation are required", invocation.ErrInvalidInvocation)
-		}
-		input, err := workflowSystemToolObjectArg(pluginMap, "input", "target.plugin")
-		if err != nil {
-			return coreworkflow.Target{}, err
-		}
-		return coreworkflow.Target{Plugin: &coreworkflow.PluginTarget{
-			PluginName: pluginName,
-			Operation:  operation,
-			Connection: workflowSystemToolStringArg(pluginMap, "connection"),
-			Instance:   workflowSystemToolStringArg(pluginMap, "instance"),
-			Input:      input,
-		}}, nil
-	}
-	agentMap, ok := workflowSystemToolMap(agentValue)
-	if !ok {
-		return coreworkflow.Target{}, fmt.Errorf("%w: target.agent must be an object", invocation.ErrInvalidInvocation)
-	}
-	if err := workflowSystemToolRejectUnknownKeys(agentMap, "target.agent", "provider", "providerName", "model", "prompt", "messages", "toolRefs", "tools", "responseSchema", "metadata", "providerOptions", "timeoutSeconds"); err != nil {
-		return coreworkflow.Target{}, err
-	}
-	toolRefsValue, ok := agentMap["toolRefs"]
-	if !ok || toolRefsValue == nil {
-		toolRefsValue = agentMap["tools"]
-	}
-	toolRefs, err := workflowSystemToolRefsFromValue(toolRefsValue)
-	if err != nil {
-		return coreworkflow.Target{}, err
-	}
-	messages, err := workflowSystemToolMessagesFromValue(agentMap["messages"])
-	if err != nil {
-		return coreworkflow.Target{}, err
-	}
-	responseSchema, err := workflowSystemToolObjectArg(agentMap, "responseSchema", "target.agent")
-	if err != nil {
-		return coreworkflow.Target{}, err
-	}
-	metadata, err := workflowSystemToolObjectArg(agentMap, "metadata", "target.agent")
-	if err != nil {
-		return coreworkflow.Target{}, err
-	}
-	providerOptions, err := workflowSystemToolObjectArg(agentMap, "providerOptions", "target.agent")
-	if err != nil {
-		return coreworkflow.Target{}, err
-	}
-	return coreworkflow.Target{Agent: &coreworkflow.AgentTarget{
-		ProviderName:    workflowSystemToolStringArg(agentMap, "provider", "providerName"),
-		Model:           workflowSystemToolStringArg(agentMap, "model"),
-		Prompt:          workflowSystemToolStringArg(agentMap, "prompt"),
-		Messages:        messages,
-		ToolRefs:        toolRefs,
-		ToolSource:      coreagent.ToolSourceModeNativeSearch,
-		ResponseSchema:  responseSchema,
-		Metadata:        metadata,
-		ProviderOptions: providerOptions,
-		TimeoutSeconds:  workflowSystemToolIntArg(agentMap, "timeoutSeconds"),
-	}}, nil
-}
-
-func workflowSystemToolRefsFromValue(value any) ([]coreagent.ToolRef, error) {
-	if value == nil {
-		return nil, nil
-	}
-	items, ok := value.([]any)
-	if !ok {
-		return nil, fmt.Errorf("%w: agent toolRefs must be an array", invocation.ErrInvalidInvocation)
-	}
-	out := make([]coreagent.ToolRef, 0, len(items))
-	for i, item := range items {
-		refMap, ok := workflowSystemToolMap(item)
-		if !ok {
-			return nil, fmt.Errorf("%w: agent toolRefs[%d] must be an object", invocation.ErrInvalidInvocation, i)
-		}
-		if err := workflowSystemToolRejectUnknownKeys(refMap, fmt.Sprintf("agent toolRefs[%d]", i), "system", "plugin", "operation", "connection", "instance", "title", "description"); err != nil {
-			return nil, err
-		}
-		out = append(out, coreagent.ToolRef{
-			System:      workflowSystemToolStringArg(refMap, "system"),
-			Plugin:      workflowSystemToolStringArg(refMap, "plugin"),
-			Operation:   workflowSystemToolStringArg(refMap, "operation"),
-			Connection:  workflowSystemToolStringArg(refMap, "connection"),
-			Instance:    workflowSystemToolStringArg(refMap, "instance"),
-			Title:       workflowSystemToolStringArg(refMap, "title"),
-			Description: workflowSystemToolStringArg(refMap, "description"),
-		})
-	}
-	return out, nil
-}
-
-func workflowSystemToolMessagesFromValue(value any) ([]coreagent.Message, error) {
-	if value == nil {
-		return nil, nil
-	}
-	items, ok := value.([]any)
-	if !ok {
-		return nil, fmt.Errorf("%w: agent messages must be an array", invocation.ErrInvalidInvocation)
-	}
-	out := make([]coreagent.Message, 0, len(items))
-	for i, item := range items {
-		messageMap, ok := workflowSystemToolMap(item)
-		if !ok {
-			return nil, fmt.Errorf("%w: agent messages[%d] must be an object", invocation.ErrInvalidInvocation, i)
-		}
-		if err := workflowSystemToolRejectUnknownKeys(messageMap, fmt.Sprintf("agent messages[%d]", i), "role", "text", "metadata"); err != nil {
-			return nil, err
-		}
-		metadata, err := workflowSystemToolObjectArg(messageMap, "metadata", fmt.Sprintf("agent messages[%d]", i))
-		if err != nil {
-			return nil, err
-		}
-		out = append(out, coreagent.Message{
-			Role:     workflowSystemToolStringArg(messageMap, "role"),
-			Text:     workflowSystemToolStringArg(messageMap, "text"),
-			Metadata: metadata,
-		})
-	}
-	return out, nil
-}
-
-func workflowSystemToolValidateCreateScope(req agentSystemToolExecutionRequest, target coreworkflow.Target) error {
-	if target.Plugin != nil {
-		if workflowSystemToolPluginTargetAllowed(*target.Plugin, req.ToolRefs, req.Tools) {
-			return nil
-		}
-		return fmt.Errorf("%w: workflow target %s.%s is outside the current agent tool scope", invocation.ErrScopeDenied, target.Plugin.PluginName, target.Plugin.Operation)
-	}
-	if target.Agent == nil {
-		return fmt.Errorf("%w: workflow target is required", invocation.ErrInvalidInvocation)
-	}
-	for i := range target.Agent.ToolRefs {
-		ref := target.Agent.ToolRefs[i]
-		if strings.TrimSpace(ref.System) != "" {
-			if err := workflowSystemToolValidateFutureSystemRef(i, ref, req); err != nil {
-				return err
-			}
-			continue
-		}
-		if strings.TrimSpace(ref.Plugin) == "" || strings.TrimSpace(ref.Plugin) == "*" || strings.TrimSpace(ref.Operation) == "" {
-			return fmt.Errorf("%w: target.agent.toolRefs[%d] must be an exact plugin operation", invocation.ErrInvalidInvocation, i)
-		}
-		if ref.CredentialMode != "" {
-			return fmt.Errorf("%w: target.agent.toolRefs[%d] credentialMode is not supported for scheduled agent targets", invocation.ErrInvalidInvocation, i)
-		}
-		pluginTarget := coreworkflow.PluginTarget{
-			PluginName: ref.Plugin,
-			Operation:  ref.Operation,
-			Connection: ref.Connection,
-			Instance:   ref.Instance,
-		}
-		if !workflowSystemToolPluginTargetAllowed(pluginTarget, req.ToolRefs, req.Tools) {
-			return fmt.Errorf("%w: target.agent.toolRefs[%d] %s.%s is outside the current agent tool scope", invocation.ErrScopeDenied, i, ref.Plugin, ref.Operation)
-		}
-	}
-	return nil
-}
-
-func workflowSystemToolValidateFutureSystemRef(index int, ref coreagent.ToolRef, req agentSystemToolExecutionRequest) error {
-	if strings.TrimSpace(ref.System) != coreagent.SystemToolWorkflow || strings.TrimSpace(ref.Operation) == "" {
-		return fmt.Errorf("%w: target.agent.toolRefs[%d] workflow system refs require an exact operation", invocation.ErrInvalidInvocation, index)
-	}
-	if strings.TrimSpace(ref.Plugin) != "" || strings.TrimSpace(ref.Connection) != "" || strings.TrimSpace(ref.Instance) != "" || ref.CredentialMode != "" {
-		return fmt.Errorf("%w: target.agent.toolRefs[%d] system refs cannot include plugin, connection, instance, or credentialMode", invocation.ErrInvalidInvocation, index)
-	}
-	for i := range req.ToolRefs {
-		if strings.TrimSpace(req.ToolRefs[i].System) == coreagent.SystemToolWorkflow && strings.TrimSpace(req.ToolRefs[i].Operation) == strings.TrimSpace(ref.Operation) {
-			return nil
-		}
-	}
-	for i := range req.Tools {
-		if strings.TrimSpace(req.Tools[i].Target.System) == coreagent.SystemToolWorkflow && strings.TrimSpace(req.Tools[i].Target.Operation) == strings.TrimSpace(ref.Operation) {
-			return nil
-		}
-	}
-	return fmt.Errorf("%w: target.agent.toolRefs[%d] workflow.%s is outside the current agent tool scope", invocation.ErrScopeDenied, index, ref.Operation)
-}
-
-func workflowSystemToolPluginTargetAllowed(target coreworkflow.PluginTarget, refs []coreagent.ToolRef, tools []coreagent.Tool) bool {
-	for i := range refs {
-		if workflowSystemToolPluginRefMatchesTarget(refs[i], target) {
-			return true
-		}
-	}
-	for i := range tools {
-		if workflowSystemToolResolvedToolMatchesTarget(tools[i], target) {
-			return true
-		}
-	}
-	return false
-}
-
-func workflowSystemToolPluginRefMatchesTarget(ref coreagent.ToolRef, target coreworkflow.PluginTarget) bool {
-	if strings.TrimSpace(ref.System) != "" || strings.TrimSpace(ref.Plugin) == "" || strings.TrimSpace(ref.Plugin) == "*" || strings.TrimSpace(ref.Operation) == "" {
-		return false
-	}
-	if ref.CredentialMode != "" {
-		return false
-	}
-	if strings.TrimSpace(ref.Plugin) != strings.TrimSpace(target.PluginName) || strings.TrimSpace(ref.Operation) != strings.TrimSpace(target.Operation) {
-		return false
-	}
-	return workflowSystemToolRefBindingMatchesTarget(ref.Connection, ref.Instance, target.Connection, target.Instance)
-}
-
-func workflowSystemToolResolvedToolMatchesTarget(tool coreagent.Tool, target coreworkflow.PluginTarget) bool {
-	if strings.TrimSpace(tool.Target.System) != "" || strings.TrimSpace(tool.Target.Plugin) == "" || strings.TrimSpace(tool.Target.Operation) == "" {
-		return false
-	}
-	if tool.Target.CredentialMode != "" {
-		return false
-	}
-	if strings.TrimSpace(tool.Target.Plugin) != strings.TrimSpace(target.PluginName) || strings.TrimSpace(tool.Target.Operation) != strings.TrimSpace(target.Operation) {
-		return false
-	}
-	return workflowSystemToolResolvedBindingMatchesTarget(tool.Target.Connection, tool.Target.Instance, target.Connection, target.Instance)
-}
-
-func workflowSystemToolRefBindingMatchesTarget(scopeConnection, scopeInstance, targetConnection, targetInstance string) bool {
-	scopeConnection = config.ResolveConnectionAlias(strings.TrimSpace(scopeConnection))
-	targetConnection = config.ResolveConnectionAlias(strings.TrimSpace(targetConnection))
-	if scopeConnection != "" && scopeConnection != targetConnection {
-		return false
-	}
-	if scopeInstance = strings.TrimSpace(scopeInstance); scopeInstance != "" && scopeInstance != strings.TrimSpace(targetInstance) {
-		return false
-	}
-	return true
-}
-
-func workflowSystemToolResolvedBindingMatchesTarget(scopeConnection, scopeInstance, targetConnection, targetInstance string) bool {
-	scopeConnection = config.ResolveConnectionAlias(strings.TrimSpace(scopeConnection))
-	targetConnection = config.ResolveConnectionAlias(strings.TrimSpace(targetConnection))
-	if scopeConnection != targetConnection {
-		return false
-	}
-	if strings.TrimSpace(scopeInstance) != strings.TrimSpace(targetInstance) {
-		return false
-	}
-	return true
-}
-
-func workflowSystemToolPermissionsForTarget(target coreworkflow.Target) []core.AccessPermission {
-	operationsByPlugin := map[string]map[string]struct{}{}
-	add := func(pluginName, operation string) {
-		pluginName = strings.TrimSpace(pluginName)
-		operation = strings.TrimSpace(operation)
-		if pluginName == "" || operation == "" {
-			return
-		}
-		if operationsByPlugin[pluginName] == nil {
-			operationsByPlugin[pluginName] = map[string]struct{}{}
-		}
-		operationsByPlugin[pluginName][operation] = struct{}{}
-	}
-	if target.Plugin != nil {
-		add(target.Plugin.PluginName, target.Plugin.Operation)
-	}
-	if target.Agent != nil {
-		for i := range target.Agent.ToolRefs {
-			ref := target.Agent.ToolRefs[i]
-			if strings.TrimSpace(ref.System) == "" {
-				add(ref.Plugin, ref.Operation)
-			}
-		}
-	}
-	if len(operationsByPlugin) == 0 {
-		return nil
-	}
-	plugins := slices.Sorted(maps.Keys(operationsByPlugin))
-	out := make([]core.AccessPermission, 0, len(plugins))
-	for _, pluginName := range plugins {
-		operations := slices.Sorted(maps.Keys(operationsByPlugin[pluginName]))
-		out = append(out, core.AccessPermission{Plugin: pluginName, Operations: operations})
-	}
-	return out
-}
-
-func workflowSystemToolScopedPrincipal(p *principal.Principal, permissions []core.AccessPermission) (*principal.Principal, error) {
-	p = principal.Canonicalized(p)
-	if p == nil || strings.TrimSpace(p.SubjectID) == "" {
-		return nil, fmt.Errorf("%w: agent execution principal is required", invocation.ErrAuthorizationDenied)
-	}
-	if len(permissions) == 0 {
-		return p, nil
-	}
-	requested := principal.CompilePermissions(permissions)
-	if requested == nil {
-		return p, nil
-	}
-	if p.TokenPermissions != nil {
-		requested = principal.IntersectPermissions(requested, p.TokenPermissions)
-		if len(requested) == 0 {
-			return nil, fmt.Errorf("%w: workflow target is outside the caller permission scope", invocation.ErrScopeDenied)
-		}
-	}
-	next := *p
-	next.TokenPermissions = requested
-	next.Scopes = principal.PermissionPlugins(requested)
-	return principal.Canonicalize(&next), nil
-}
-
-func workflowSystemToolScheduleInfo(schedule *workflowmanager.ManagedSchedule) map[string]any {
-	value := map[string]any{}
-	if schedule == nil {
-		return value
-	}
-	if providerName := strings.TrimSpace(schedule.ProviderName); providerName != "" {
-		value["provider"] = providerName
-	}
-	if schedule.Schedule != nil {
-		coreSchedule := schedule.Schedule
-		value["id"] = coreSchedule.ID
-		value["cron"] = coreSchedule.Cron
-		value["timezone"] = coreSchedule.Timezone
-		value["paused"] = coreSchedule.Paused
-		value["target"] = workflowSystemToolTargetInfo(coreSchedule.Target)
-		workflowSystemToolPutTime(value, "createdAt", coreSchedule.CreatedAt)
-		workflowSystemToolPutTime(value, "updatedAt", coreSchedule.UpdatedAt)
-		workflowSystemToolPutTime(value, "nextRunAt", coreSchedule.NextRunAt)
-	}
-	return value
 }
 
 func workflowSystemRunInfo(run *workflowmanager.ManagedRun) map[string]any {
@@ -817,52 +311,6 @@ func workflowSystemToolStringArg(args map[string]any, keys ...string) string {
 		}
 	}
 	return ""
-}
-
-func workflowSystemToolBoolArg(args map[string]any, key string) bool {
-	value, ok := args[key]
-	if !ok {
-		return false
-	}
-	result, _ := value.(bool)
-	return result
-}
-
-func workflowSystemToolIntArg(args map[string]any, key string) int {
-	value, ok := args[key]
-	if !ok {
-		return 0
-	}
-	switch v := value.(type) {
-	case int:
-		return v
-	case int64:
-		return int(v)
-	case float64:
-		return int(v)
-	default:
-		return 0
-	}
-}
-
-func workflowSystemToolMap(value any) (map[string]any, bool) {
-	if value == nil {
-		return nil, false
-	}
-	out, ok := value.(map[string]any)
-	return out, ok
-}
-
-func workflowSystemToolObjectArg(args map[string]any, key, path string) (map[string]any, error) {
-	value, ok := args[key]
-	if !ok || value == nil {
-		return nil, nil
-	}
-	typed, ok := workflowSystemToolMap(value)
-	if !ok {
-		return nil, fmt.Errorf("%w: %s.%s must be an object", invocation.ErrInvalidInvocation, path, key)
-	}
-	return maps.Clone(typed), nil
 }
 
 func workflowSystemToolMapDeepClone(value map[string]any) map[string]any {
