+++ gestaltd/internal/server/routes_user.go
@@ -9,27 +9,27 @@ func (s *Server) mountAuthenticatedRoutes(r chi.Router) {
 		r.Get("/integrations", s.listIntegrations)
 		r.Delete("/integrations/{name}", s.disconnectIntegration)
 
-		r.Get("/workflow/schedules", s.listGlobalWorkflowSchedules)
-		r.Post("/workflow/schedules", s.createWorkflowSchedule)
+		r.Get("/workflow/schedules", s.workflowScheduleCRUDRemoved)
+		r.Post("/workflow/schedules", s.workflowScheduleCRUDRemoved)
 		r.Route("/workflow/schedules", func(r chi.Router) {
-			r.Get("/", s.listGlobalWorkflowSchedules)
-			r.Post("/", s.createWorkflowSchedule)
-			r.Get("/{scheduleID}", s.getGlobalWorkflowSchedule)
-			r.Put("/{scheduleID}", s.updateGlobalWorkflowSchedule)
-			r.Delete("/{scheduleID}", s.deleteGlobalWorkflowSchedule)
-			r.Post("/{scheduleID}/pause", s.pauseGlobalWorkflowSchedule)
-			r.Post("/{scheduleID}/resume", s.resumeGlobalWorkflowSchedule)
+			r.Get("/", s.workflowScheduleCRUDRemoved)
+			r.Post("/", s.workflowScheduleCRUDRemoved)
+			r.Get("/{scheduleID}", s.workflowScheduleCRUDRemoved)
+			r.Put("/{scheduleID}", s.workflowScheduleCRUDRemoved)
+			r.Delete("/{scheduleID}", s.workflowScheduleCRUDRemoved)
+			r.Post("/{scheduleID}/pause", s.workflowScheduleCRUDRemoved)
+			r.Post("/{scheduleID}/resume", s.workflowScheduleCRUDRemoved)
 		})
-		r.Get("/workflow/event-triggers", s.listGlobalWorkflowEventTriggers)
-		r.Post("/workflow/event-triggers", s.createGlobalWorkflowEventTrigger)
+		r.Get("/workflow/event-triggers", s.workflowEventTriggerCRUDRemoved)
+		r.Post("/workflow/event-triggers", s.workflowEventTriggerCRUDRemoved)
 		r.Route("/workflow/event-triggers", func(r chi.Router) {
-			r.Get("/", s.listGlobalWorkflowEventTriggers)
-			r.Post("/", s.createGlobalWorkflowEventTrigger)
-			r.Get("/{triggerID}", s.getGlobalWorkflowEventTrigger)
-			r.Put("/{triggerID}", s.updateGlobalWorkflowEventTrigger)
-			r.Delete("/{triggerID}", s.deleteGlobalWorkflowEventTrigger)
-			r.Post("/{triggerID}/pause", s.pauseGlobalWorkflowEventTrigger)
-			r.Post("/{triggerID}/resume", s.resumeGlobalWorkflowEventTrigger)
+			r.Get("/", s.workflowEventTriggerCRUDRemoved)
+			r.Post("/", s.workflowEventTriggerCRUDRemoved)
+			r.Get("/{triggerID}", s.workflowEventTriggerCRUDRemoved)
+			r.Put("/{triggerID}", s.workflowEventTriggerCRUDRemoved)
+			r.Delete("/{triggerID}", s.workflowEventTriggerCRUDRemoved)
+			r.Post("/{triggerID}/pause", s.workflowEventTriggerCRUDRemoved)
+			r.Post("/{triggerID}/resume", s.workflowEventTriggerCRUDRemoved)
 		})
 		r.Post("/workflow/events", s.publishWorkflowEvent)
 		r.Get("/workflow/runs", s.listGlobalWorkflowRuns)
