+++ gestaltd/internal/server/test_helpers_test.go
@@ -1,0 +1,12 @@
+package server_test
+
+func cloneMap(src map[string]any) map[string]any {
+	if src == nil {
+		return nil
+	}
+	dst := make(map[string]any, len(src))
+	for key, value := range src {
+		dst[key] = value
+	}
+	return dst
+}
